Logging info/debug to stdout, others to stderr.
Using a simple but ugly way to have logrus emit panic/fatal/error/warn
log statements to os.Stderr whilst info and debug statements go to
stdout. Doing the ugly thing for stderr (which has more levels) since I
expect to log info/debug more often so this should be more "efficient".

Logrus calls hooks before logging to console so I'm adding a hook that
duplicates the Logger struct value, changes Out, and sets it back to
entry. Logger is a pointer so I must make a copy to avoid race
conditions, however entry.log() is a non-pointer reciever so it should
be safe to change a field value (Logger).